Rename create_params.py module
